Add power button
