+++ lib/domain/scooter_battery.dart
