+++ lib/home_screen.dart
@@ -9,6 +9,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:unustasis/control_screen.dart';
 import 'package:unustasis/domain/icomoon.dart';
 import 'package:unustasis/driving_screen.dart';
+import 'package:unustasis/interfaces/phone/scooter_power_button.dart';
 import 'package:unustasis/onboarding_screen.dart';
 import 'package:unustasis/scooter_service.dart';
 import 'package:unustasis/domain/scooter_state.dart';
@@ -263,38 +264,8 @@ class _HomeScreenState extends State<HomeScreen> {
                             : null,
                       ),
                     ),
-                    Expanded(
-                      child: ScooterPowerButton(
-                          action: _scooterState != null &&
-                                  _scooterState!.isReadyForLockChange
-                              ? (_scooterState!.isOn
-                                  ? () {
-                                      try {
-                                        widget.scooterService.lock();
-                                      } catch (e) {
-                                        if (e
-                                            .toString()
-                                            .contains("SEAT_OPEN")) {
-                                          showSeatWarning();
-                                        } else {
-                                          Fluttertoast.showToast(
-                                              msg: e.toString());
-                                        }
-                                      }
-                                    }
-                                  : (_scooterState == ScooterState.standby
-                                      ? widget.scooterService.unlock
-                                      : widget.scooterService.wakeUpAndUnlock))
-                              : null,
-                          icon: _scooterState != null && _scooterState!.isOn
-                              ? Icons.lock_open
-                              : Icons.lock_outline,
-                          label: _scooterState != null && _scooterState!.isOn
-                              ? FlutterI18n.translate(
-                                  context, "home_lock_button")
-                              : FlutterI18n.translate(
-                                  context, "home_unlock_button")),
-                    ),
+                    ScooterPowerButtonContainer(
+                        _scooterState, widget.scooterService),
                     Expanded(
                       child: ScooterActionButton(
                           onPressed: !_scanning
@@ -409,82 +380,6 @@ class _HomeScreenState extends State<HomeScreen> {
     } else {
       widget.scooterService.optionalAuth = true;
     }
-  }
-}
-
-class ScooterPowerButton extends StatefulWidget {
-  const ScooterPowerButton({
-    super.key,
-    required void Function()? action,
-    Widget? child,
-    required IconData icon,
-    required String label,
-  })  : _action = action,
-        _icon = icon,
-        _label = label;
-
-  final void Function()? _action;
-  final String _label;
-  final IconData _icon;
-
-  @override
-  State<ScooterPowerButton> createState() => _ScooterPowerButtonState();
-}
-
-class _ScooterPowerButtonState extends State<ScooterPowerButton> {
-  bool loading = false;
-
-  @override
-  Widget build(BuildContext context) {
-    Color mainColor = widget._action == null
-        ? Theme.of(context).colorScheme.onSurface.withOpacity(0.2)
-        : Theme.of(context).colorScheme.onSurface;
-    return Column(
-      children: [
-        ElevatedButton(
-          style: ElevatedButton.styleFrom(
-            padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 24),
-            backgroundColor: mainColor,
-          ),
-          onPressed: () {
-            Fluttertoast.showToast(msg: widget._label);
-          },
-          onLongPress: widget._action == null
-              ? null
-              : () {
-                  setState(() {
-                    loading = true;
-                  });
-                  widget._action!();
-                  Future.delayed(const Duration(seconds: 5), () {
-                    setState(() {
-                      loading = false;
-                    });
-                  });
-                },
-          child: loading
-              ? SizedBox(
-                  height: 16,
-                  width: 16,
-                  child: CircularProgressIndicator(
-                    color: Theme.of(context).colorScheme.background,
-                  ),
-                )
-              : Icon(
-                  widget._icon,
-                  color: Theme.of(context).colorScheme.background,
-                ),
-        ),
-        const SizedBox(height: 16),
-        Text(
-          widget._label,
-          style: TextStyle(
-            color: mainColor,
-          ),
-          textAlign: TextAlign.center,
-        ),
-      ],
-    );
   }
 }
